NISP-1656: Refactored IDAProvider into VerifyProvider
